Add Image column to Staffs table
A new migration file named `20240813071206_add_image_column_with_staff_table.cs` has been added. This migration introduces a new column named `Image` to the `Staffs` table in the database. The `Image` column is of type `nvarchar(max)` and is nullable. The `Up` method in the migration adds the `Image` column to the `Staffs` table, while the `Down` method removes it, allowing for rollback if needed.