Add Image and Name properties to ContextModelSnapshot
Added two new string properties, Image and Name, to the
ContextModelSnapshot.cs file in the HotelProject.DataAccessLayer.Migrations
namespace. These properties are configured to be stored as nvarchar(max)
in the database, allowing the model to store additional information
such as an image URL or file path and a name.