Remove unused columns from Booking table and entity class
Updated the database schema and corresponding entity class to remove
the `CreatedDate`, `TotalPrice`, and `UpdatedDate` columns from the
`Bookings` table. Added a new migration class `updatebookingtable`
in `20240902094506_update-booking-table.Designer.cs` to define the
structure of the `Booking` table. Implemented the `Up` method in
`20240902094506_update-booking-table.cs` to remove the columns and
the `Down` method to add them back if needed. Updated the `Booking`
entity class (`Booking.cs`) to reflect these changes.